Split Pandera asset check factory into its own module.
